requirements.txt
Flask
llama-index-llms-gemini==0.2.0
llama-index
google-generativeai
llama-index-core
llama-index-embeddings-gemini
python-dotenv
SQLAlchemy
PyMySQL
